+++ src/main/java/system/customer/responseDTO/CustomerResponseDTO.java
@@ -4,7 +4,7 @@ import system.customer.model.CustomerModel;
 
 import java.util.UUID;
 
-public record CustomerResponseDTO(UUID id, int document, int numberPhone, String name, String email) {
+public record CustomerResponseDTO(UUID id, long document, long numberPhone, String name, String email) {
     public CustomerResponseDTO(CustomerModel customerModel) {
         this(
             customerModel.getId(),
